+++ lib/vero.rb
@@ -1,54 +1,24 @@
 # frozen_string_literal: true
 
+require "base64"
+require "json"
 require "rest-client"
+require "zeitwerk"
 
-module Vero
-  autoload :Config, "vero/config"
-  autoload :App, "vero/app"
-  autoload :Context, "vero/context"
-  autoload :APIContext, "vero/context/api"
-  autoload :Trackable, "vero/trackable"
-  autoload :DSL, "vero/dsl"
-  autoload :Sender, "vero/sender"
-  autoload :SuckerPunchWorker, "vero/senders/sucker_punch"
-  autoload :ResqueWorker, "vero/senders/resque"
-  autoload :SidekiqWorker, "vero/senders/sidekiq"
-
-  module Api
-    module Workers
-      autoload :BaseAPI, "vero/api/base_api"
-
-      module Events
-        autoload :TrackAPI, "vero/api/events/track_api"
-      end
-
-      module Users
-        autoload :TrackAPI, "vero/api/users/track_api"
-        autoload :EditAPI, "vero/api/users/edit_api"
-        autoload :EditTagsAPI, "vero/api/users/edit_tags_api"
-        autoload :UnsubscribeAPI, "vero/api/users/unsubscribe_api"
-        autoload :ResubscribeAPI, "vero/api/users/resubscribe_api"
-        autoload :ReidentifyAPI, "vero/api/users/reidentify_api"
-        autoload :DeleteAPI, "vero/api/users/delete_api"
-      end
-    end
-
-    autoload :Events, "vero/api"
-    autoload :Users, "vero/api"
-  end
-
-  module Senders
-    autoload :Base, "vero/senders/base"
-    autoload :DelayedJob, "vero/senders/delayed_job"
-    autoload :Resque, "vero/senders/resque"
-    autoload :Sidekiq, "vero/senders/sidekiq"
-    autoload :Invalid, "vero/senders/invalid"
-    autoload :SuckerPunch, "vero/senders/sucker_punch"
-  end
-
-  module Utility
-    autoload :Logger, "vero/utility/logger"
-  end
-end
+loader = Zeitwerk::Loader.for_gem
+loader.inflector.inflect(
+  "api_context" => "APIContext",
+  "dsl" => "DSL",
+  "base_api" => "BaseAPI",
+  "track_api" => "TrackAPI",
+  "delete_api" => "DeleteAPI",
+  "edit_api" => "EditAPI",
+  "edit_tags_api" => "EditTagsAPI",
+  "reidentify_api" => "ReidentifyAPI",
+  "resubscribe_api" => "ResubscribeAPI",
+  "unsubscribe_api" => "UnsubscribeAPI"
+)
+loader.ignore("#{__dir__}/generators")
+loader.setup
 
 require "vero/railtie" if defined?(Rails)
