+++ lib/vero/api/base.rb
@@ -1,0 +1,27 @@
+# frozen_string_literal: true
+
+class Vero::Api::Base
+  attr_accessor :context
+
+  def initialize(context)
+    self.context = context
+  end
+
+  def config
+    context.config
+  end
+
+  def run_api(api_klass, options)
+    return if config.disabled
+
+    validate_configured!
+    options.merge!(config.request_params)
+    Vero::Sender.send(api_klass, config.async, config.domain, options)
+  end
+
+  protected
+
+  def validate_configured!
+    raise "You must configure the 'vero' gem. Visit https://github.com/getvero/vero for more details." unless config.configured?
+  end
+end
