--- lib/vero/api/events/track_api.rb
@@ -1,24 +1,0 @@
-# frozen_string_literal: true
-
-module Vero
-  module Api
-    module Workers
-      module Events
-        class TrackAPI < BaseAPI
-          def url
-            "#{@domain}/api/v2/events/track.json"
-          end
-
-          def request
-            RestClient.post(url, request_params_as_json, request_content_type)
-          end
-
-          def validate!
-            raise ArgumentError, "Missing :event_name" if options[:event_name].to_s.strip.empty?
-            raise ArgumentError, ":data must be either nil or a Hash" unless options[:data].nil? || options[:data].is_a?(Hash)
-          end
-        end
-      end
-    end
-  end
-end
