--- lib/vero/api/users/reidentify_api.rb
@@ -1,24 +1,0 @@
-# frozen_string_literal: true
-
-module Vero
-  module Api
-    module Workers
-      module Users
-        class ReidentifyAPI < BaseAPI
-          def url
-            "#{@domain}/api/v2/users/reidentify.json"
-          end
-
-          def request
-            RestClient.put(url, request_params_as_json, request_content_type)
-          end
-
-          def validate!
-            raise ArgumentError, "Missing :id" if options[:id].to_s.strip.empty?
-            raise ArgumentError, "Missing :new_id" if options[:new_id].to_s.strip.empty?
-          end
-        end
-      end
-    end
-  end
-end
