+++ lib/vero/app.rb
@@ -1,27 +1,25 @@
 # frozen_string_literal: true
 
-module Vero
-  class App
-    include Vero::Utility::Logger
+class Vero::App
+  include Vero::Utility::Logger
 
-    def self.default_context
-      @@default_context ||= Context.new # rubocop:disable Style/ClassVars
-    end
+  def self.default_context
+    @@default_context ||= Vero::Context.new # rubocop:disable Style/ClassVars
+  end
 
-    def self.init(&block)
-      default_context.configure(&block) if block
-    end
+  def self.init(&block)
+    default_context.configure(&block) if block
+  end
 
-    def self.reset!
-      default_context.reset!
-    end
+  def self.reset!
+    default_context.reset!
+  end
 
-    def self.disable_requests!
-      default_context.disable_requests!
-    end
+  def self.disable_requests!
+    default_context.disable_requests!
+  end
 
-    def self.configured?
-      default_context.configured?
-    end
+  def self.configured?
+    default_context.configured?
   end
 end
