+++ lib/vero/trackable.rb
@@ -1,8 +1,5 @@
 # frozen_string_literal: true
 
-require "vero/trackable/base"
-require "vero/trackable/interface"
-
 module Vero
   module Trackable
     include Base
