+++ lib/vero/utility/logger.rb
@@ -1,29 +1,25 @@
 # frozen_string_literal: true
 
-module Vero
-  module Utility
-    module Logger
-      def self.included(base)
-        base.extend(ClassMethods)
+module Vero::Utility::Logger
+  def self.included(base)
+    base.extend(ClassMethods)
+  end
+
+  module ClassMethods
+    def log(object, message)
+      return unless Vero::App.default_context.config.logging && !defined?(RSpec)
+
+      message = "#{object.class.name}: #{message}"
+
+      if (logger = self.logger)
+        logger.info(message)
+      else
+        puts(message)
       end
+    end
 
-      module ClassMethods
-        def log(object, message)
-          return unless Vero::App.default_context.config.logging && !defined?(RSpec)
-
-          message = "#{object.class.name}: #{message}"
-
-          if (logger = self.logger)
-            logger.info(message)
-          else
-            puts(message)
-          end
-        end
-
-        def logger
-          Rails.logger if defined?(Rails)
-        end
-      end
+    def logger
+      Rails.logger if defined?(Rails)
     end
   end
 end
