+++ spec/spec_helper.rb
@@ -2,8 +2,8 @@
 
 require "rubygems"
 require "bundler/setup"
+
 require "vero"
-require "json"
 
 Dir[File.expand_path("support/**/*.rb", __dir__)].sort.each { |f| require f }
