+++ vero.gemspec
@@ -22,4 +22,5 @@ Gem::Specification.new do |spec|
 
   spec.add_dependency "json"
   spec.add_dependency "rest-client"
+  spec.add_dependency "zeitwerk"
 end
